e-com-backend/requirements.txt
asgiref==3.4.1
asyncio==3.4.3
CacheControl==0.12.10
cachelib==0.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudinary==1.28.0
colorama==0.4.4
crypto==1.4.1
dnspython==2.1.0
email-validator==1.1.3
firebase==3.0.1
firebase-admin==5.1.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-Session==0.4.0
Flask-WTF==1.0.0
gcloud==0.18.3
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jws==0.1.3
MarkupSafe==2.0.1
mongoengine==0.23.1
msgpack==1.0.3
Naked==0.1.31
oauth2client==4.1.3
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.12.0
pymongo==3.12.2
pyparsing==3.0.6
python-dotenv==0.19.2
python-http-client==3.3.4
python-jwt==2.0.1
PyYAML==6.0
razorpay==1.2.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
sendgrid==6.9.2
shellescape==3.8.1
six==1.16.0
sseclient==0.0.27
starkbank-ecdsa==2.0.3
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
